test: move auth step 4 prose test into the prose test file
